TrainBackends: Split the @module_type decorator out of module.py
This has three benefits:

1. Better code encapsulation
2. Avoids circular dependencies with module backend setup
3. Easier to remove if/when we decide to do away with module subtypes
